+++ src/assets/data/BulletinData.jsx
@@ -1,0 +1,32 @@
+ const bulletinData = [
+  {
+    date: "24 Jun 2025, 10:04 AM",
+    title: "Exclusion of F&O contracts on 8 securities from August 29, 2025",
+  },
+  {
+    date: "17 Jun 2025, 05:09 PM",
+    title: "Revision in expiry day of Index and Stock derivatives contracts",
+  },
+  {
+    date: "01 Jul 2025, 03:09 PM",
+    title: "Quarterly Settlement of Funds - July 2025",
+  },
+  {
+    date: "01 Jul 2025, 08:10 AM",
+    title: "Offer for sale (OFS) – June 2025",
+  },
+  {
+    date: "01 Jul 2025, 08:00 AM",
+    title: "Surveillance measure on scrips - July 2025",
+  },
+  {
+    date: "01 Jul 2025, 08:00 AM",
+    title: "Current Takeovers and Delisting – July 2025",
+  },
+  {
+    date: "01 Jul 2025, 07:41 AM",
+    title: "Latest Intraday leverages and Square-off timings",
+  }
+];
+
+export default bulletinData ;
