+++ src/assets/data/CircularData.jsx
@@ -1,0 +1,34 @@
+
+ const circular = [
+  { date: "26 Jun 2025", title: "Rights issue of Mirc Electronics Ltd", exchange: "BSE" },
+  { date: "26 Jun 2025", title: "Merger of Umang Dairies Ltd", exchange: "BSE" },
+  { date: "26 Jun 2025", title: "Rights issue of Bodhtree Consulting Ltd", exchange: "BSE" },
+  { date: "26 Jun 2025", title: "Adjustment of F&O contracts of Concor due to bonus", exchange: "NSE" },
+  { date: "25 Jun 2025", title: "Change in the symbol of Mahaveer Infoway Ltd", exchange: "BSE" },
+  { date: "25 Jun 2025", title: "Change in name and symbol of Vandana Knitwear Ltd", exchange: "BSE" },
+  { date: "25 Jun 2025", title: "Change in name and symbol of S R G Securities Finance Ltd", exchange: "BSE" },
+  { date: "24 Jun 2025", title: "Rights issue of GVP Infotech Ltd", exchange: "BSE" },
+  { date: "24 Jun 2025", title: "Exclusion of F&O contract on 8 securities", exchange: "NSE" },
+  { date: "24 Jun 2025", title: "Rights entitlements of Capital Trust Ltd.", exchange: "BSE" },
+  { date: "23 Jun 2025", title: "Bonus issue of Investment & Precision Castings Ltd", exchange: "BSE" },
+  { date: "23 Jun 2025", title: "Rights issue of Infibeam Avenues Ltd", exchange: "BSE" },
+  { date: "23 Jun 2025", title: "Change in name and symbol of Adinath Exim Resources Ltd", exchange: "BSE" },
+  { date: "23 Jun 2025", title: "Change in name and symbol of KBC Global Ltd", exchange: "NSE" },
+  { date: "20 Jun 2025", title: "Rights entitlements of 7NR Retail Ltd", exchange: "BSE" },
+  { date: "20 Jun 2025", title: "Rights entitlements of IFL Enterprises Ltd", exchange: "BSE" },
+  { date: "20 Jun 2025", title: "Stock split of Laddu Gopal Online Services Ltd", exchange: "BSE" },
+  { date: "20 Jun 2025", title: "Change in the name and symbol of Sat Industries Ltd", exchange: "BSE" },
+  { date: "19 Jun 2025", title: "Rights entitlements of Som Datt Finance Corporation Ltd", exchange: "BSE" },
+  { date: "18 Jun 2025", title: "Rights entitlements of Ethos Ltd", exchange: "BSE" },
+  { date: "18 Jun 2025", title: "Merger of Inox Wind Energy ltd", exchange: "BSE" },
+  { date: "18 Jun 2025", title: "Adjustment of F&O contracts of PNB due to dividend", exchange: "NSE" },
+  { date: "18 Jun 2025", title: "Change in the name and symbol of Kavveri Telecom Products Ltd", exchange: "BSE" },
+  { date: "18 Jun 2025", title: "Stock split of Elitecon International Ltd", exchange: "BSE" },
+  { date: "17 Jun 2025", title: "Rights Entitlements of Accent Microcell Ltd", exchange: "NSE" },
+  { date: "17 Jun 2025", title: "Rights entitlements of Sylph Technologies Ltd", exchange: "BSE" },
+  { date: "16 Jun 2025", title: "Adjustment of F&O contracts of BAJAJ-AUTO due to dividend", exchange: "NSE" },
+  { date: "16 Jun 2025", title: "Adjustment of F&O contracts of TECHM due to dividend", exchange: "NSE" },
+  { date: "16 Jun 2025", title: "Reduction of equity share capital of Shri Krishna Prasadam Ltd", exchange: "BSE" },
+];
+
+export default circular;
