+++ src/assets/data/HolidaysData.jsx
@@ -1,0 +1,77 @@
+
+const holidaysData = [
+  {
+    day: "Wednesday",
+    date: "26 Feb 2025",
+    holiday: "Maha Shivaratri",
+    exchanges: ["NSE", "BSE"],
+  },
+  {
+    day: "Thursday",
+    date: "10 Apr 2025",
+    holiday: "Mahavir Jayanti",
+    exchanges: ["NSE", "BSE"],
+  },
+  {
+    day: "Monday",
+    date: "14 Apr 2025",
+    holiday: "Dr. Baba Saheb Ambedkar Jayanti",
+    exchanges: ["NSE", "BSE"],
+  },
+  {
+    day: "Friday",
+    date: "18 Apr 2025",
+    holiday: "Good Friday",
+    exchanges: ["NSE", "BSE", "MCX"],
+  },
+  {
+    day: "Thursday",
+    date: "01 May 2025",
+    holiday: "Maharashtra Day",
+    exchanges: ["NSE", "BSE"],
+  },
+  {
+    day: "Friday",
+    date: "15 Aug 2025",
+    holiday: "Independence Day",
+    exchanges: ["NSE", "BSE", "MCX"],
+  },
+  {
+    day: "Wednesday",
+    date: "27 Aug 2025",
+    holiday: "Ganesh Chaturthi",
+    exchanges: ["NSE", "BSE"],
+  },
+  {
+    day: "Thursday",
+    date: "02 Oct 2025",
+    holiday: "Mahatma Gandhi Jayanti",
+    exchanges: ["NSE", "BSE"],
+  },
+  {
+    day: "Tuesday",
+    date: "21 Oct 2025",
+    holiday: "Diwali-Laxmi Pujan (Muhurat trading session)",
+    exchanges: ["NSE", "BSE", "MCX"],
+  },
+  {
+    day: "Wednesday",
+    date: "22 Oct 2025",
+    holiday: "Diwali-Balipratipada",
+    exchanges: ["NSE", "BSE"],
+  },
+  {
+    day: "Wednesday",
+    date: "05 Nov 2025",
+    holiday: "Gurunanak Jayanti",
+    exchanges: ["NSE", "BSE"],
+  },
+  {
+    day: "Thursday",
+    date: "25 Dec 2025",
+    holiday: "Christmas",
+    exchanges: ["NSE", "BSE", "MCX"],
+  },
+];
+
+export default holidaysData;
