+++ src/assets/data/InvestoreContent.jsx
