+++ src/assets/data/McxHolidaysData.jsx
@@ -1,0 +1,40 @@
+
+const mcxHolidaysData = [
+  {
+    day: "Sunday",
+    date: "26 Jan 2025",
+    holiday: "Republic Day",
+    exchange: "MCX",
+  },
+  {
+    day: "Friday",
+    date: "18 Apr 2025",
+    holiday: "Good Friday",
+    exchange: "MCX",
+  },
+  {
+    day: "Friday",
+    date: "15 Aug 2025",
+    holiday: "Independence Day",
+    exchange: "MCX",
+  },
+  {
+    day: "Thursday",
+    date: "02 Oct 2025",
+    holiday: "Mahatma Gandhi Jayanti",
+    exchange: "MCX",
+  },
+  {
+    day: "Tuesday",
+    date: "21 Oct 2025",
+    holiday: "Diwali-Laxmi Pujan (Muhurat trading session)",
+    exchange: "MCX",
+  },
+  {
+    day: "Thursday",
+    date: "25 Dec 2025",
+    holiday: "Christmas",
+    exchange: "MCX",
+  },
+];
+export default mcxHolidaysData;
