+++ src/assets/data/SettlementHolidaysData.jsx
@@ -1,0 +1,28 @@
+
+const settlementHolidaysData = [
+  {
+    day: "Wednesday",
+    date: "19 Feb 2025",
+    holiday: "Chhatrapati Shivaji Maharaj Jayanti",
+    type: "SETTLEMENT",
+  },
+  {
+    day: "Tuesday",
+    date: "01 Apr 2025",
+    holiday: "Annual Bank closing",
+    type: "SETTLEMENT",
+  },
+  {
+    day: "Monday",
+    date: "12 May 2025",
+    holiday: "Buddha Pournima",
+    type: "SETTLEMENT",
+  },
+  {
+    day: "Friday",
+    date: "05 Sep 2025",
+    holiday: "Id-E-Milad",
+    type: "SETTLEMENT",
+  },
+];
+export default  settlementHolidaysData;
